// CLIENT/src/store/actions/actionsTypes.ts
type actions = {
    UPDATE_ITEM: string;
    SORTED_BY_LIKES: string;
    SORTED_BY_TAG_ECOLOGIA: string;
    SORTED_BY_TAG_MEDICINA: string;
    SORTED_BY_TAG_TRANSPORTE: string;
    SORTED_BY_TAG_INCLUSION: string;
    SORTED_BY_TAG_ECONOMIA: string;
    SORTED_BY_TIME: string;
};

export const ACTIONS: actions = {
    UPDATE_ITEM: 'UPDATE_ITEM',
    SORTED_BY_LIKES: 'SORTED_BY_LIKES',
    SORTED_BY_TIME: 'SORTED_BY_TIME',
    SORTED_BY_TAG_ECOLOGIA: 'SORTED_BY_TAG_ECOLOGIA',
    SORTED_BY_TAG_MEDICINA: 'SORTED_BY_TAG_MEDICINA',
    SORTED_BY_TAG_TRANSPORTE: 'SORTED_BY_TAG_TRANSPORTE',
    SORTED_BY_TAG_INCLUSION: 'SORTED_BY_TAG_INCLUSION',
    SORTED_BY_TAG_ECONOMIA: 'SORTED_BY_TAG_ECONOMIA',
};